Fix pass button showing a 'selected' card
In the prop-drilling version, there was a secret extra prop on the Card
component: whether selectedCard was passed as the real value in state or
simply left undefined. This sort of indirectly controlled whether the
card was able to style itself as selected or not. The pass button was
relying on this behavior prior to the refactor.

This commit adds an explicit prop to the Card component to control this
behavior.